requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
dnspython==2.1.0
envyaml==1.8.210417
idna==2.10
kafka-python==2.0.2
multidict==5.1.0
PyYAML==5.4.1
requests==2.25.1
typing-extensions==3.7.4.3
urllib3==1.26.4
yarl==1.6.3
